pyproject.toml
[project]
name = "ai-resume-matcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.5",
    "click>=8.3.1",
    "fastapi>=0.123.8",
    "google-generativeai>=0.8.5",
    "graphiti-core>=0.24.1",
    "httpx>=0.28.1",
    "langchain>=1.1.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-groq>=1.1.0",
    "langchain-openai>=1.1.0",
    "mem0ai>=1.0.1",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.4.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.5",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.52.0",
    "structlog>=25.5.0",
    "urllib3>=2.5.0",
    "uvicorn[standard]>=0.38.0",
]
